fix: code 리뷰 반영
- AuctionDocumentReindexJob 정리 예외가 job에 기록되도록 수정
- 옛 인덱스가 없을 경우 삭제 호출을 하지 않기

+++ src/main/java/com/example/auction/domain/auction/search/service/AuctionDocumentReindexService.java
@@ -36,16 +36,16 @@ public class AuctionDocumentReindexService {
         log.info("[AuctionDocumentReindexService] elasticsearch document reindexing 작업 시작");
 
         LocalDateTime beganTime = LocalDateTime.now();
-        LocalDateTime monthAgo = beganTime.minusDays(30);
-
-        // 오래된 작업 내역 들을 삭제
-        helper.deleteOldReindexJobs(monthAgo);
 
         String newIndexName = AuctionDocumentUtil.getNewIndexName(beganTime);
 
         AuctionDocumentReindexJob job = helper.saveJob(AuctionDocumentReindexJob.of(newIndexName));
 
         try {
+            LocalDateTime monthAgo = beganTime.minusDays(30);
+            // 오래된 작업 내역 들을 삭제
+            helper.deleteOldReindexJobs(monthAgo);
+
             // 백업을 할 새 index 생성
             IndexCoordinates target = IndexCoordinates.of(newIndexName);
             var indexOps = elasticsearch.indexOps(AuctionDocument.class);
@@ -100,7 +100,9 @@ public class AuctionDocumentReindexService {
             }
 
             List<String> oldIndexes = helper.pointAliasAtNewIndex(AuctionDocumentUtil.ALIAS_NAME, newIndexName);
-            elasticsearch.indexOps(IndexCoordinates.of(oldIndexes.toArray(new String[0]))).delete();
+            if (!oldIndexes.isEmpty()) {
+                elasticsearch.indexOps(IndexCoordinates.of(oldIndexes.toArray(new String[0]))).delete();
+            }
 
             job.updateJobStatus(AuctionDocumentReindexJobStatus.DONE);
             helper.saveJob(job);
